requirements.txt
absl-py==0.11.0
gunicorn==20.0.4
astunparse==1.6.3
certifi==2020.12.5
Flask==1.1.2
gast==0.3.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
Markdown==3.3.3
MarkupSafe==1.1.1
numpy==1.19.3
oauthlib==3.1.0
opt-einsum==3.3.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.1
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
tensorboard==2.4
tensorflow-cpu
tensorflow-estimator<2.6.0
imageai==2.1.6
opencv-python-headless
pillow==7.0.0
pandas
pytesseract